+++ app/src/androidTest/java/org/simple/clinic/medicalhistory/MedicalHistoryRepositoryAndroidTest.kt
@@ -1,21 +1,20 @@
 package org.simple.clinic.medicalhistory
 
 import com.google.common.truth.Truth.assertThat
-import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.simple.clinic.TestClinicApp
-import org.simple.sharedTestCode.TestData
 import org.simple.clinic.medicalhistory.Answer.No
 import org.simple.clinic.medicalhistory.Answer.Unanswered
 import org.simple.clinic.medicalhistory.Answer.Yes
 import org.simple.clinic.patient.SyncStatus
 import org.simple.clinic.rules.LocalAuthenticationRule
 import org.simple.clinic.rules.SaveDatabaseRule
-import org.simple.sharedTestCode.util.Rules
 import org.simple.clinic.util.UtcClock
+import org.simple.sharedTestCode.TestData
+import org.simple.sharedTestCode.util.Rules
 import java.time.Instant
 import java.time.temporal.ChronoUnit.DAYS
 import java.util.UUID
@@ -53,7 +52,9 @@ class MedicalHistoryRepositoryAndroidTest {
         hasHadHeartAttack = Yes,
         hasHadStroke = Yes,
         hasHadKidneyDisease = Yes,
-        hasDiabetes = No)
+        hasDiabetes = No,
+        isSmoking = No,
+    )
 
     repository.save(
         uuid = UUID.fromString("d33a3dfc-3da9-43a9-a543-095232c55597"),
@@ -67,6 +68,7 @@ class MedicalHistoryRepositoryAndroidTest {
     assertThat(savedHistory.hasHadStroke).isEqualTo(Yes)
     assertThat(savedHistory.hasHadKidneyDisease).isEqualTo(Yes)
     assertThat(savedHistory.diagnosedWithDiabetes).isEqualTo(No)
+    assertThat(savedHistory.isSmoking).isEqualTo(No)
     assertThat(savedHistory.syncStatus).isEqualTo(SyncStatus.PENDING)
   }
 
@@ -119,6 +121,7 @@ class MedicalHistoryRepositoryAndroidTest {
     assertThat(emptyHistory.hasHadStroke).isEqualTo(Unanswered)
     assertThat(emptyHistory.hasHadKidneyDisease).isEqualTo(Unanswered)
     assertThat(emptyHistory.diagnosedWithDiabetes).isEqualTo(Unanswered)
+    assertThat(emptyHistory.isSmoking).isEqualTo(Unanswered)
     assertThat(emptyHistory.syncStatus).isEqualTo(SyncStatus.DONE)
   }
 
@@ -178,6 +181,7 @@ class MedicalHistoryRepositoryAndroidTest {
     assertThat(emptyHistory.hasHadKidneyDisease).isEqualTo(Unanswered)
     assertThat(emptyHistory.diagnosedWithHypertension).isEqualTo(Unanswered)
     assertThat(emptyHistory.diagnosedWithDiabetes).isEqualTo(Unanswered)
+    assertThat(emptyHistory.isSmoking).isEqualTo(Unanswered)
     assertThat(emptyHistory.syncStatus).isEqualTo(SyncStatus.DONE)
   }
 }