+++ app/src/androidTest/java/org/simple/clinic/storage/migrations/Migration116AndroidTest.kt
@@ -1,0 +1,43 @@
+package org.simple.clinic.storage.migrations
+
+import org.junit.Test
+import org.simple.clinic.assertColumns
+
+class Migration116AndroidTest : BaseDatabaseMigrationTest(115, 116) {
+
+  @Test
+  fun isSmoking_should_be_added_to_medical_history_table() {
+    before.assertColumns("MedicalHistory", setOf(
+        "uuid",
+        "patientUuid",
+        "diagnosedWithHypertension",
+        "isOnHypertensionTreatment",
+        "isOnDiabetesTreatment",
+        "hasHadHeartAttack",
+        "hasHadStroke",
+        "hasHadKidneyDisease",
+        "hasDiabetes",
+        "syncStatus",
+        "createdAt",
+        "updatedAt",
+        "deletedAt"
+    ))
+
+    after.assertColumns("MedicalHistory", setOf(
+        "uuid",
+        "patientUuid",
+        "diagnosedWithHypertension",
+        "isOnHypertensionTreatment",
+        "isOnDiabetesTreatment",
+        "hasHadHeartAttack",
+        "hasHadStroke",
+        "hasHadKidneyDisease",
+        "hasDiabetes",
+        "isSmoking",
+        "syncStatus",
+        "createdAt",
+        "updatedAt",
+        "deletedAt"
+    ))
+  }
+}